Simplify admin & user sign in spec helpers
Many of them are only used in one file,
so move them to the file that they're used in.

Motivation: there was confusion between `sign_in_admin`,
which walked through the user interface to sign in an admin,
and `login_admin`, which used Devise & sessions to set the user
directly.