Remove import logs
See https://trello.com/c/g985Z0ar

- remove a bunch of unused code from ImportJob

The import logic was rather complicated by keeping track of the import
log, which was displayed after the import to the user.

After talking with Deb Sherwin, we realized that the import log
isn't really being used by the user. If anything, it's more useful to
devs.

Only devs care what the state of the data was at the time of import.
Users care about the state of the data *currently*.

Because of that, we're going to tear it out of the interface, and just
store the logs in the standard Rails logger. I might look into using
Honeybadger or something to keep up on errors in addition to straight
logging.

Fix style violations